# Add varlen attention