### Description

This PR adds a Triton `varlen_attention` implementation.

### Testing

Please select all that apply.

- [ ] Existing unit tests
- [x] Unit tests added by this PR
- [ ] Other (please explain)
- [ ] This PR is not tested

#### Test instructions

```bash
CONCH_ENABLE_VLLM=1 pytest tests/varlen_attention_test.py
```

```bash
864 passed in 87.61s (0:01:27)
```

#### Platforms

Please select all hardware platforms that this PR was tested on.

- [x] Nvidia GPU
- [ ] AMD GPU
- [ ] Other (please explain)

#### Benchmarks

**A10**

```bash
CONCH_ENABLE_VLLM=1 python benchmarks/varlen_attention_benchmark.py
```

```bash
Results matched with atol=0.001 :)
Parameters: {'head_dim': 256, 'seq_len': 1024, 'cache_block_size': 32, 'batch_size': 10, 'num_query_heads': 8, 'num_kv_heads': 4, 'causal': False, 'pure_decode': False}
Triton: num_iterations=100, min=2.516 ms, max=2.686 ms, mean=2.551 ms, median=2.544 ms
Baseline: num_iterations=100, min=0.784 ms, max=0.889 ms, mean=0.854 ms, median=0.881 ms
```

```bash
CONCH_ENABLE_VLLM=1 python benchmarks/varlen_attention_benchmark.py --causal
```

```bash
Results matched with atol=0.001 :)
Parameters: {'head_dim': 256, 'seq_len': 1024, 'cache_block_size': 32, 'batch_size': 10, 'num_query_heads': 8, 'num_kv_heads': 4, 'causal': True, 'pure_decode': False}
Triton: num_iterations=100, min=1.740 ms, max=1.888 ms, mean=1.798 ms, median=1.791 ms
Baseline: num_iterations=100, min=0.485 ms, max=0.526 ms, mean=0.507 ms, median=0.509 ms
```